step3 - refactor package 조정
