+++ src/test/java/study/racing/RacingInputTest.java
@@ -2,7 +2,6 @@ package study.racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import study.racing.view.RacingInput;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
